requirements.txt
rel==0.4.9.19
websocket-client[optional]==1.8.0
aiohttp==3.10.0
aiohttp_middlewares==2.3.0
fastapi==0.111.0
pydantic==2.7.2
pydantic-settings==2.3.0
uvicorn==0.30.0

einops==0.8.0
matplotlib==3.8.0
numpy==1.26.4
numba==0.59.0
pandas==2.1.4
scikit-learn==1.2.2
scipy==1.11.4
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0

ffmpeg-python==0.2.0
audioread==3.0.1
librosa==0.10.2.post1
pooch==1.8.2
pyaudio==0.2.14
soundfile==0.12.1
soxr==0.4.0